+++ app/src/main/java/org/simple/clinic/instantsearch/InstantSearchEvent.kt
@@ -3,6 +3,11 @@ package org.simple.clinic.instantsearch
 import org.simple.clinic.bp.assignbppassport.BlankBpPassportResult
 import org.simple.clinic.facility.Facility
 import org.simple.clinic.patient.PatientSearchResult
+import org.simple.clinic.patient.businessid.Identifier
+import org.simple.clinic.scanid.EnteredShortCode
+import org.simple.clinic.scanid.PatientFound
+import org.simple.clinic.scanid.PatientNotFound
+import org.simple.clinic.scanid.ScanResult
 import org.simple.clinic.widgets.UiEvent
 import java.util.UUID
 
@@ -33,4 +38,24 @@ object RegisterNewPatientClicked : InstantSearchEvent() {
 data class BlankBpPassportResultReceived(val blankBpPassportResult: BlankBpPassportResult) : InstantSearchEvent() {
 
   override val analyticsName: String = "Instant Search : Blank BP Passport Result Received"
+}
+
+sealed class BpPassportScanned : InstantSearchEvent() {
+
+  companion object {
+
+    fun fromResult(scanResult: ScanResult): BpPassportScanned {
+      return when (scanResult) {
+        is EnteredShortCode -> ByShortCode(scanResult.shortCode)
+        is PatientFound -> ByPatientFound(scanResult.patientId)
+        is PatientNotFound -> ByPatientNotFound(scanResult.identifier)
+      }
+    }
+  }
+
+  data class ByPatientFound(val patientId: UUID) : BpPassportScanned()
+
+  data class ByPatientNotFound(val identifier: Identifier) : BpPassportScanned()
+
+  data class ByShortCode(val shortCode: String) : BpPassportScanned()
 }