+++ app/src/test/java/org/simple/clinic/instantsearch/InstantSearchEffectHandlerTest.kt
@@ -308,4 +308,19 @@ class InstantSearchEffectHandlerTest {
     verify(uiActions).showKeyboard()
     verifyNoMoreInteractions(uiActions)
   }
+
+  @Test
+  fun `when open short code search effect is received, then open short code search`() {
+    // given
+    val shortCode = "123 4567"
+
+    // when
+    testCase.dispatch(OpenShortCodeSearchScreen(shortCode))
+
+    // then
+    testCase.assertNoOutgoingEvents()
+
+    verify(uiActions).openShortCodeSearchScreen(shortCode)
+    verifyNoMoreInteractions(uiActions)
+  }
 }